requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
groq==0.11.0
Werkzeug==3.0.1
httpx==0.27.0